♻️ Move command.go to command package
